Fix blocked weights with FQ strategies
when a weight reached 1.00005 for instance, it was never increased
due to the comparison on floats.
Also, the weight was no more increased when the step became too small
compared to the minimum weight. As we are not using integers but
floats, the step must remain proportional to the weight.